+++ Domain/Interfaces/IUserFinanceSystem/IUserFinanceSystem.cs
@@ -5,7 +5,7 @@ namespace Domain.Interfaces.IUserFinanceSystem
 {
     public interface IUserFinanceSystem : GenericInterface<UserFinanceSystem>
     {
-        Task<IEnumerable<UserFinanceSystem>> GetSystemUsers(long idSystem);
+        Task<IEnumerable<UserFinanceSystem>> GetSystemUsers(int idSystem);
         Task RemoveSystemUsers(IEnumerable<UserFinanceSystem> users);
         Task <UserFinanceSystem> GetUserByEmail(string userEmail);
     }
